# ahm92/app/resources/seng365_event_site_api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SENG365 2021: Meetup Site API",
    "version": "1.0.0",
    "description": "This specification documents the API that must be implemented for Assignment 1.\nThe API provides all of the server-side functionality necessary in order to implement the user stories for Assignment 2."
  },
  "servers": [
    {
      "url": "http://seng365-apitest.csse.canterbury.ac.nz:4001/api/v1",
      "description": "reference server"
    }
  ],
...
  "paths": {
    "/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "View events.",
        "description": "By default, the events should be sorted by date of the events.\nThe list should be filtered, then sorted, then paginated using `startIndex` and `count`. ",
        "parameters": [
          {
            "in": "query",
            "name": "startIndex",
            "description": "Number of items to skip before returning results.",
            "schema": {
              "$ref": "#/components/schemas/EventSearchRequest/properties/startIndex"
            }
          },
          {
            "in": "query",
            "name": "count",
            "description": "Number of items to include in results.",
            "schema": {
              "$ref": "#/components/schemas/EventSearchRequest/properties/count"
            }
          },
          {
            "in": "query",
            "name": "q",
            "description": "Only include events that have the search term within their title or description.",
            "schema": {
              "$ref": "#/components/schemas/EventSearchRequest/properties/q"
            }
          },
          {
            "in": "query",
            "name": "categoryIds",
            "description": "Only include events of this category (id).\n\nOnly IDs that map to categories are allowed.",
            "schema": {
              "$ref": "#/components/schemas/EventSearchRequest/properties/categoryIds"
            }
          },
          {
            "in": "query",
            "name": "organizerId",
            "description": "Only include events that have the given user (id) as their organizer.",
            "schema": {
              "$ref": "#/components/schemas/EventSearchRequest/properties/organizerId"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "description": "Sort the events by the given property, according to the following rules:\n* ALPHABETICAL_ASC: alphabetically by title, A - Z\n* ALPHABETICAL_DESC: alphabetically by title, Z - A\n* DATE_ASC: date, from earliest to latest\n* DATE_DESC: date, from latest to earliest\n* ATTENDEES_ASC: the number of signed up (approved) attendees, from least to most\n* ATTENDEES_DESC: the number of signed up (approved) attendees, from most to least\n* CAPACITY_ASC: the capacity from least to most\n* CAPACITY_DESC: the capacity from most to least\n\nNote: how ties (e.g. two events with 10 attendees each) are decided is undocumented behaviour, although it should be consistent each time.",
            "schema": {
              "$ref": "#/components/schemas/EventSearchRequest/properties/sortBy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOverview"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Add a new event.",
        "description": "The elements of the category ids array `categoryIds` must each reference an existing category.\n\n`title`, `description`, and `categoryIds` are required, other properties are optional. An event without `capacity` defined or set to `null` means there is no limit on the number of attendees.\n\nThe event `date` must be in the future.",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "CreateEventRequest",
                "properties": {
                  "title": {
                    "$ref": "#/components/schemas/EventOverview/properties/title"
                  },
                  "description": {
                    "$ref": "#/components/schemas/Event/properties/description"
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Category/properties/categoryId"
                    }
                  },
                  "date": {
                    "$ref": "#/components/schemas/DateTimeInput"
                  },
                  "isOnline": {
                    "$ref": "#/components/schemas/Event/properties/isOnline"
                  },
                  "url": {
                    "$ref": "#/components/schemas/Event/properties/url"
                  },
                  "venue": {
                    "$ref": "#/components/schemas/Event/properties/venue"
                  },
                  "capacity": {
                    "$ref": "#/components/schemas/EventOverview/properties/capacity"
                  },
                  "requiresAttendanceControl": {
                    "$ref": "#/components/schemas/Event/properties/requiresAttendanceControl"
                  },
                  "fee": {
                    "$ref": "#/components/schemas/Event/properties/fee"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "categoryIds"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "eventId": {
                      "$ref": "#/components/schemas/EventOverview/properties/eventId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events/{id}": {
      "parameters": [
        {
          "name": "id",
          "description": "ID of the event.",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/EventOverview/properties/eventId"
          }
        }
      ],
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Retrieve detailed information about an event.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "events"
        ],
        "summary": "Change an event's details.",
        "description": "Only accessible to the organizer of the event. If updated, the `categoryIds` must each reference an existing category.\n\nNot accessible after the event has happened (i.e. the date is in the past).",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "ChangeEventDetailsRequest",
                "properties": {
                  "title": {
                    "$ref": "#/components/schemas/EventOverview/properties/title"
                  },
                  "description": {
                    "$ref": "#/components/schemas/Event/properties/description"
                  },
                  "categoryIds": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Category/properties/categoryId"
                    }
                  },
                  "date": {
                    "$ref": "#/components/schemas/DateTimeInput"
                  },
                  "isOnline": {
                    "$ref": "#/components/schemas/Event/properties/isOnline"
                  },
                  "url": {
                    "$ref": "#/components/schemas/Event/properties/url"
                  },
                  "venue": {
                    "$ref": "#/components/schemas/Event/properties/venue"
                  },
                  "capacity": {
                    "$ref": "#/components/schemas/EventOverview/properties/capacity"
                  },
                  "requiresAttendanceControl": {
                    "$ref": "#/components/schemas/Event/properties/requiresAttendanceControl"
                  },
                  "fee": {
                    "$ref": "#/components/schemas/Event/properties/fee"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Delete an event.",
        "description": "Only accessible to the organizer of the event. All attendees for the event will also be deleted.",
        "security": [
          {
            "UserToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/events/{id}/image": {
      "parameters": [
        {
          "name": "id",
          "description": "ID of the event.",
          "in": "path",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/EventOverview/properties/eventId"
          }
        }
      ],
      "get": {
        "tags": [
          "events.images"
        ],
        "summary": "Retrieve an event's hero image.",
        "description": "The response MIME type will be either `image/png`, `image/jpeg`, or `image/gif`, depending on the file type of the image being retrieved.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryImageFile"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "events.images"
        ],
        "summary": "Set an event's hero image.",
        "description": "Only accessible to the organizer of the event. If the event already has a hero image, the current hero image will be replaced with it, and a 200 OK response will be sent. If not, a 201 Created response will be sent.\n\nIf the file is a `.png`, the request will be sent with `Content-Type: image/png`. If the file is a `.jpeg` or `.jpg`, the request will be sent with `Content-Type: image/jpeg`. If the file is a `.gif`, the request will be sent with `Content-Type: image/gif`.\n\nThe request body should contain the raw binary content of the photo, i.e. it should not be in `multipart/form-data` format or contain any file metadata. ",
        "security": [
          {
            "UserToken": []
          }
        ],
        "requestBody": {
          "content": {
            "image/png": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            },
            "image/jpeg": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            },
            "image/gif": {
              "schema": {
                "$ref": "#/components/schemas/BinaryImageFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Event": {
        "title": "Event",
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/EventOverview"
          }
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "Pizza party to celebrate the nice weather."
          },
          "organizerId": {
            "$ref": "#/components/schemas/FullUser/properties/userId"
          },
          "date": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DateTimeOutput"
              }
            ],
            "nullable": true
          },
          "isOnline": {
            "type": "boolean",
            "example": false,
            "nullable": true
          },
          "url": {
            "type": "string",
            "example": "http://example.com",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "example": "Town Hall",
            "nullable": true
          },
          "requiresAttendanceControl": {
            "type": "boolean",
            "example": true,
            "nullable": true
          },
          "fee": {
            "type": "number",
            "example": 10,
            "nullable": true
          }
        }
      },
      "EventOverview": {
        "title": "EventOverview",
        "type": "object",
        "properties": {
          "eventId": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "example": "Pizza party"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category/properties/categoryId"
            }
          },
          "organizerFirstName": {
            "$ref": "#/components/schemas/User/properties/firstName"
          },
          "organizerLastName": {
            "$ref": "#/components/schemas/User/properties/lastName"
          },
          "numAcceptedAttendees": {
            "type": "integer",
            "minimum": 0,
            "example": 37,
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "minimum": 1,
            "example": 100,
            "nullable": true
          }
        }
      },
...
      "DateTimeInput": {
        "title": "DateTimeInput",
        "description": "A datetime expressed with no timezone, as either a date (yyyy-MM-dd) or date and time (yyyy-MM-dd hh:mm:ss.sss)",
        "type": "string",
        "example": "2012-04-23 18:25:43.511"
      },
      "DateTimeOutput": {
        "title": "DateTimeOutput",
        "description": "A UTC datetime expressed in ISO8601 format (yyyy-MM-ddThh:mm:ss.sssZ)",
        "type": "string",
        "example": "2012-04-23T18:25:43.511Z"
      },
      "BinaryImageFile": {
        "type": "string",
        "format": "binary",
        "description": "The raw content of the PNG, JPEG, or GIF image file.",
        "example": "(raw content of PNG, JPEG, or GIF image file)"
      },
...
    },
    "securitySchemes": {
      "UserToken": {
        "type": "apiKey",
        "name": "X-Authorization",
        "in": "header"
      }
    }
  }
}